moved subaward tests to new file
